Change to renderSidebar hook
By changing to activate on the first call of the renderSidebar hook, we can bypass a split second moment on first loading in when the sidebar shows as the incorrect width. I haven't noticed any issues being caused by this change, though I admit my testing has not been exhaustive.